Move ResolvedVisitor.h to headers
---
