+++ frontend/test/resolution/testVarArgs.cpp
@@ -19,16 +19,15 @@
 
 #include "test-resolution.h"
 
-#include "ResolvedVisitor.h"
-
 #include "chpl/framework/ErrorWriter.h"
 #include "chpl/parsing/parsing-queries.h"
+#include "chpl/resolution/ResolvedVisitor.h"
 #include "chpl/resolution/resolution-queries.h"
 #include "chpl/resolution/scope-queries.h"
 #include "chpl/types/all-types.h"
-#include "chpl/uast/IntentList.h"
-#include "chpl/uast/Identifier.h"
 #include "chpl/uast/For.h"
+#include "chpl/uast/Identifier.h"
+#include "chpl/uast/IntentList.h"
 #include "chpl/uast/Module.h"
 #include "chpl/uast/Record.h"
 #include "chpl/uast/Variable.h"
